fix: revert the negative boundary logic changes in the simple differencing script
While it would be nice to keep the logic between the simple and advanced differencing queries consistent, this introduced a bug with 1 day intervals in the left set. The advanced version can filter this scenario out because it maintains the order of the boundary events and filters out rows with an adjusted start date after the adjusted end date. The alternative to reverting the simple script would be to update it to also use the fixed ordering instead of GROUP BY in the outer most query.